+++ packages/byte-bank/src/app/api/account/transaction/route.ts
@@ -1,7 +1,5 @@
-"use server";
-
-import { NextRequest, NextResponse } from "next/server";
-import { TransactionService } from '$/server/services/transaction.service';
+import {NextRequest, NextResponse} from "next/server";
+import {TransactionService} from "$/server/services/transaction.service";
 
 export async function GET(req: NextRequest) {
   try {
@@ -16,7 +14,7 @@ export async function GET(req: NextRequest) {
     }
 
     const transactionService = new TransactionService();
-    const transaction = await transactionService.getTransactionById(userId, transactionId);
+    const transaction = await transactionService.getTransactionById(transactionId, userId, req);
 
     if (!transaction) {
       return NextResponse.json({error: 'Transaction not found', status: 404});
@@ -24,6 +22,32 @@ export async function GET(req: NextRequest) {
 
     return NextResponse.json({transaction: transaction, status: 200});
 
+  } catch (e) {
+    console.error(e);
+    let errorMessage = "Server error.";
+    if (e instanceof Error) {
+      errorMessage = e.message;
+    }
+    return NextResponse.json(
+      {message: errorMessage},
+      {
+        status: 500,
+      }
+    );
+  }
+}
+
+export async function POST(req: NextRequest) {
+  try {
+    const transactionService = new TransactionService();
+
+    const body = await req.text();
+    const { type, amount, user } = JSON.parse(body);
+
+    const newTransaction =
+      await transactionService.createTransaction({type, amount, user}, req);
+
+    return NextResponse.json({transaction: newTransaction, status: 201});
   } catch (e) {
     console.error(e);
     let errorMessage = "Server error.";
